Modify minCost function to use vector<int>
Changed function signature to accept vector of integers instead of long long.

+++ GFG/Minimum Cost of ropes.cpp
@@ -1,8 +1,10 @@
+//Approach-1 :
+//The key to minimizing the cost is to always connect the smallest ropes first, because connecting the biggest ones early will lead to a higher total cost. We can use a min-heap (a special kind of list that always give the smallest element first) to help with this.
 class Solution {
   public:
     // Function to return the minimum cost of connecting the ropes.
-    long long minCost(vector<long long>& arr) {
-        // Your code here
+    int minCost(vector<int>& arr) {
+        // code here
         priority_queue<long long, vector<long long>, greater<long long>>pq;
       
         for(int i=0;i<arr.size();i++){
